requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
pandas==1.5.3
numpy==1.24.3
ta==0.10.1
gunicorn==20.1.0
python-dotenv==1.0.0
setuptools>=65.5.0
pip>=23.0
requests==2.31.0